remove reset password of ppconsole
